refactor: move server to its own file
